Glich language, change the way local variables are handled.
Zero level variables are variables declared in the main script as opposed to in functions and commands.

The lifetime of zero level variables are controlled by the 'mark' system. The 'clear' statement is now redundant and will be removed.

When running a function or command, the previous locals are swapped out by replacing the m_store member of Glich and restoring it afterwards.

A consequence of these changes is the 'mark' statement can't be used in a function or command, namely a non zero level.